Fix failing render tests and use {} instead of Set()
Matches the pull request submitted to the main repo, which catches some cases we hadn't noticed, like there being a region specified on body, or no body present at all. Also removes use of `Set()` which has a name clash with a `Set()` object defined in the library, and is an es6 feature.